invoke_image_caption/requirements.txt
pillow==10.3.0
pillow_heif==0.22.0